Split the room loop into media and turn
The room module carried the whole agent side of an interview in one
file, and its own header asked that a split follow its section list
rather than a line count. Two of those sections are now their own
modules: the buffers, pumps and codecs that carry audio and video in
both directions, and the state an interview is a sequence of.

The procedures that end a turn stayed with the loop even though they
read as turn logic. Every one of them borrows the loop's event context,
and a module that has to reach back into its parent for the type it
operates on is the same code in two files rather than a seam. What is
left of the turn module depends only on the agent module, and the media
module reaches up for a single predicate.

No behavior changes. Every item is the one that moved, visibility is
widened only where the parent names the item, and the tests that touch
one module moved into it. The ones spanning output audio and turn
cancellation stayed, because that coupling is real and lives in the
loop.
